fix: correctly hydrate windows
this fixes a hydration issue where
scratch buffers were being populated
as they are the last window open,
and being set as the active window.

now we close these scratch buffers and
also hack neotree to properly open a window
on the left side of the screen.

the neotree hack is pretty jank, but it
does in fact work after much struggle.

Fixes #81